requirements.txt
flask==3.1.2
imutils==0.5.4
keras
mediapipe==0.10.7
opencv-python
tensorflow
numpy
pillow==12.0.0
requests==2.32.5
twilio==9.9.0
python-dotenv
